Refactoring tracing code - extracing to common and splitting to http and gRPC
Unexporting methods that are not required to be exported
